refactored everything into components in components/codenames
