+++ src/components/codenames/GameData.js
@@ -1,0 +1,27 @@
+export const initialWords = [
+    { word: "COPPER", color: "blue" },
+    { word: "RING", color: "blue" },
+    { word: "FAIR", color: "red" },
+    { word: "SUPERHERO", color: "red" },
+    { word: "SLIP", color: "neutral" },
+    { word: "IVORY", color: "neutral" },
+    { word: "LITTER", color: "neutral" },
+    { word: "WASHER", color: "red" },
+    { word: "POINT", color: "red" },
+    { word: "COMPOUND", color: "neutral" },
+    { word: "OLIVE", color: "neutral" },
+    { word: "AIR", color: "black" }, // Assassin word
+    { word: "CYCLE", color: "neutral" },
+    { word: "VET", color: "blue" },
+    { word: "OLYMPUS", color: "blue" },
+    { word: "PIRATE", color: "neutral" },
+    { word: "MINE", color: "neutral" },
+    { word: "PASTA", color: "red" },
+    { word: "DRUG", color: "blue" },
+    { word: "DUCK", color: "neutral" },
+    { word: "ENGLAND", color: "neutral" },
+    { word: "OIL", color: "neutral" },
+    { word: "LEAD", color: "red" },
+    { word: "PUMPKIN", color: "blue" },
+    { word: "CHINA", color: "neutral" },
+]
